Refactored CSharp Code generator into sub-namespace. Also added docs
