+++ src/typechat.app/PluginProgramValidator.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using Microsoft.SemanticKernel.SkillDefinition;
-using Microsoft.TypeChat;
+using Microsoft.TypeChat.CSharp;
 using Microsoft.TypeChat.Schema;
 
 namespace Microsoft.TypeChat;
